+++ Adapters/Smaato/Target.swift
@@ -1,0 +1,4 @@
+//
+//  Target.swift
+//  cas-ios-spm
+//