+++ Package.swift
@@ -3,98 +3,40 @@
 
 import PackageDescription
 
-// Some ad frameworks, such as IronSource and UnityAds, do not provide official support for
-// Swift Package Manager (SPM).
-// Instead, we define custom Package.swift configurations using the official .zip archives
-// provided for CocoaPods integration.
-// Below are the CocoaPods links used to configure the binaryTargets.
-
 let baseTarget = "CASBaseResources"
-let baseBinary = "CASBase"
+let baseBinary = "CleverAdsSolutions"
 
 enum CAS: String, CaseIterable {
     case Exchange
     case CrossPromo
-
-    /// https://github.com/CocoaPods/Specs/blob/master/Specs/7/7/b/IronSourceSDK/
     case IronSource
-
-    /// https://github.com/CocoaPods/Specs/blob/master/Specs/2/e/8/UnityAds/
     case UnityAds
-
-    /// https://github.com/CocoaPods/Specs/tree/master/Specs/6/7/d/HyprMX/
     case HyprMX
-
-    /// https://github.com/Kidoz-SDK/kidoz-sdk-swift-package
     case Kidoz
-
-    /// https://github.com/Prado-SDK/prado-sdk-swift-package
     case Prado
-
-    /// https://github.com/AppLovin/AppLovin-MAX-Swift-Package
     case AppLovin
-
-    /// https://github.com/googleads/swift-package-manager-google-mobile-ads
     case GoogleAds
-
-    /// https://github.com/Vungle/VungleAdsSDK-SwiftPackageManager
     case LiftoffMonetize
-
-    /// https://github.com/CocoaPods/Specs/blob/master/Specs/2/1/5/FBAudienceNetwork/
     case AudienceNetwork
-
-    /// https://github.com/StartApp-SDK/StartAppSDK-SwiftPackage
     case StartIO
-
-    /// https://github.com/Mintegral-official/MintegralAdSDK-Swift-Package
     case Mintegral
-
-    /// https://github.com/CocoaPods/Specs/tree/master/Specs/1/7/3/Fyber_Marketplace_SDK/
     case DTExchange
-
-    /// https://github.com/CocoaPods/Specs/tree/master/Specs/7/8/1/InMobiSDK/
     case InMobi
-
-    /// https://github.com/yandexmobile/yandex-ads-sdk-ios/blob/master/Package.swift
-    /// https://github.com/divkit/divkit-ios-facade
-    /// https://github.com/appmetrica/appmetrica-sdk-ios
     case YangoAds
-
-    /// https://github.com/CocoaPods/Specs/tree/master/Specs/8/a/a/YsoNetworkSDK/
     case YsoNetwork
-
-    /// https://github.com/bytedance/AdsGlobalPackage
-    /// https://github.com/bytedance/Bytedance-UnionAD/blob/master/Ads-Global/Ads-Global.podspec
     case Pangle
-
-    /// https://github.com/cloudadrd/zMaticooPodSpec/blob/main/zMaticoo.podspec
-    //case Maticoo
-
-    /// https://github.com/PubMatic/OpenWrapSDK-Swift-Package
     case PubMatic
-
-    /// https://github.com/Ogury/ogury-sdk-spm
     case Ogury
-    /// https://github.com/CocoaPods/Specs/blob/master/Specs/5/3/e/ChartboostSDK/
     case Chartboost
-
-    /// https://github.com/vervegroup/hybid-ios-spm-sdk
     case Verve
-
-    /// https://github.com/CocoaPods/Specs/tree/master/Specs/a/5/5/Bigo/
-    //case Bigo
-
-    /// https://github.com/displayio/DIOSDK
     case DisplayIO
-
-    /// Not supported:  Many separate frameworks
-    //case Smaato
-
-    /// Not supported: Has no compiled frameworks
+    case Smaato
+    case Moloco
+    //case Bigo
+    //case Bidease
+    //case Maticoo
+    //case Monetrix
     //case SuperAwesome
-
-    /// Not supported:  Many separate frameworks
-    //case Madex
 
     var product: Product {
         return Product.library(name: "CASMediation" + rawValue, targets: [target])
@@ -125,20 +67,29 @@ let package = Package(
     platforms: [.iOS(.v13)],
     products: CAS.allProducts,
     dependencies: [
-        .package(url: "https://github.com/googleads/swift-package-manager-google-mobile-ads", exact: "13.6.0"),
-        .package(url: "https://github.com/Vungle/VungleAdsSDK-SwiftPackageManager", exact: "7.7.4"),
-        .package(url: "https://github.com/Kidoz-SDK/kidoz-sdk-swift-package", exact: "10.1.5"),
-        .package(url: "https://github.com/Mintegral-official/MintegralAdSDK-Swift-Package", exact: "8.1.5"),
-        .package(url: "https://github.com/AppLovin/AppLovin-MAX-Swift-Package", exact: "13.6.3"),
-        .package(url: "https://github.com/bytedance/AdsGlobalPackage", exact: "8.1.1-release.0"),
-        .package(url: "https://github.com/appmetrica/appmetrica-sdk-ios", .upToNextMinor(from: "6.3.0")),
-        .package(url: "https://github.com/StartApp-SDK/StartAppSDK-SwiftPackage", exact: "4.13.0"),
-        .package(url: "https://github.com/JunGroupProductions/HyprMX-SDK-SPM", exact: "6.4.6"),
-        .package(url: "https://github.com/Ogury/ogury-sdk-spm", exact: "5.2.3"),
-        .package(url: "https://github.com/Prado-SDK/prado-sdk-swift-package", exact: "10.1.5"),
-        .package(url: "https://github.com/vervegroup/hybid-ios-spm-sdk", exact: "3.8.1"),
-        .package(url: "https://github.com/PubMatic/OpenWrapSDK-Swift-Package", exact: "5.1.1"),
-        .package(url: "https://github.com/displayio/DIOSDK", exact: "4.7.4"),
+        .package(url: "https://github.com/ironsource-mobile/LevelPlay-Swift-Package.git", exact: "9.5.0"),
+        .package(url: "https://github.com/Unity-Technologies/Unity-Ads-Swift-Package.git", exact: "4.19.0"),
+        .package(url: "https://github.com/googleads/swift-package-manager-google-mobile-ads.git", exact: "13.7.0"),
+        .package(url: "https://github.com/Vungle/VungleAdsSDK-SwiftPackageManager.git", exact: "7.7.6"),
+        .package(url: "https://github.com/InMobi/InMobiSDK-Swift-Package.git", exact: "11.4.0"),
+        .package(url: "https://github.com/ChartBoost/chartboost-monetization-ios-sdk.git", exact: "9.13.0"),
+        .package(url: "https://github.com/inner-active/DTExchangeSDK-iOS-SPM.git", exact: "8.4.9"),
+        .package(url: "https://github.com/Kidoz-SDK/kidoz-sdk-swift-package.git", exact: "10.1.5"),
+        .package(url: "https://github.com/Mintegral-official/MintegralAdSDK-Swift-Package.git", exact: "8.1.6"),
+        .package(url: "https://github.com/AppLovin/AppLovin-MAX-Swift-Package.git", exact: "13.6.3"),
+        .package(url: "https://github.com/bytedance/AdsGlobalPackage", exact: "8.2.0-release.8"),
+        .package(url: "https://github.com/yandexmobile/yandex-ads-sdk-ios.git", exact: "8.3.0"),
+        .package(url: "https://github.com/facebook/FBAudienceNetwork.git", exact: "6.22.0"),
+        .package(url: "https://bitbucket.org/ysocorp/ysonetwork-ios-sdk.git", exact: "1.2.1"),
+        .package(url: "https://github.com/StartApp-SDK/StartAppSDK-SwiftPackage.git", exact: "4.14.0"),
+        .package(url: "https://github.com/JunGroupProductions/HyprMX-SDK-SPM.git", exact: "6.4.6"),
+        .package(url: "https://github.com/Ogury/ogury-sdk-spm.git", exact: "5.2.3"),
+        .package(url: "https://github.com/Prado-SDK/prado-sdk-swift-package.git", exact: "10.1.5"),
+        .package(url: "https://github.com/vervegroup/Smaato-ios-sdk-standalone.git", exact: "23.2.0"),
+        .package(url: "https://github.com/vervegroup/hybid-ios-spm-sdk.git", exact: "3.9.0"),
+        .package(url: "https://github.com/PubMatic/OpenWrapSDK-Swift-Package.git", exact: "5.2.0"),
+        .package(url: "https://github.com/displayio/DIOSDK.git", exact: "4.7.5"),
+        .package(url: "https://github.com/moloco/moloco-sdk-ios-spm.git", exact: "4.9.0"),
     ],
 
     targets: [
@@ -161,12 +112,13 @@ let package = Package(
                 .linkedFramework("SwiftUI"),
                 .linkedFramework("UIKit"),
                 .linkedFramework("WebKit"),
+                .linkedFramework("StoreKit"),
             ]
         ),
         .binaryTarget(
             name: baseBinary,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.0-beta1/CleverAdsSolutions-4.8.0-beta1.zip",
-            checksum: "450f9a637971733c61958096f4ab0aa10d3efe5b90cddddad4fdadda997b67b3"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CleverAdsSolutions-4.8.1-beta1.zip",
+            checksum: "3ed0e4f71d73b87284fa5580a10990d7f598a5d50a4f72ee9c71fe36550d0202"
         ),
 
         // MARK: - IronSource
@@ -174,44 +126,16 @@ let package = Package(
         .target(
             name: CAS.IronSource.target,
             dependencies: [
-                .target(name: CAS.IronSource.binarySDK),
-                .target(name: "IronSourceAdQualitySPM"),
                 .target(name: CAS.IronSource.binaryAdapter),
                 .target(name: baseTarget),
+                .product(name: "UnityMediationSDK", package: "LevelPlay-Swift-Package"),
             ],
             path: CAS.IronSource.pathAdapter,
-            linkerSettings: [
-                .linkedFramework("AdSupport"),
-                .linkedFramework("AudioToolbox"),
-                .linkedFramework("AVFoundation"),
-                .linkedFramework("CFNetwork"),
-                .linkedFramework("CoreGraphics"),
-                .linkedFramework("CoreMedia"),
-                .linkedFramework("CoreTelephony"),
-                .linkedFramework("CoreVideo"),
-                .linkedFramework("Foundation"),
-                .linkedFramework("MobileCoreServices"),
-                .linkedFramework("QuartzCore"),
-                .linkedFramework("Security"),
-                .linkedFramework("StoreKit"),
-                .linkedFramework("SystemConfiguration"),
-                .linkedLibrary("z"),
-            ]
-        ),
-        .binaryTarget(
-            name: CAS.IronSource.binarySDK,
-            url: "https://github.com/ironsource-mobile/iOS-sdk/raw/master/9.5.0/IronSource9.5.0.zip",
-            checksum: "925ec91a408809da6021b8cb2f7abb6b463521872ec8952cdf29ed636e65aa3b"
-        ),
-        .binaryTarget(
-            name: "IronSourceAdQualitySPM",
-            url: "https://github.com/ironsource-mobile/iOS-adqualitysdk/releases/download/9.7.0/IronSourceAdQualitySDK-ios-v9.7.0.zip",
-            checksum: "5962af8d75db7156cfc4fcb7c7c5b006c236cfa5287518367c4c481bad1f4c3d"
         ),
         .binaryTarget(
             name: CAS.IronSource.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.0-beta1/CASMediationIronSource-9.5.0.0.zip",
-            checksum: "006d6a2d026c9845d46e4546563e79c66ee18be7030080110def19eeeddbce67"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationIronSource-9.5.0.1.zip",
+            checksum: "37f72151786691c2ac3cef4ba7d05305da7a1c1e979a774d6e494139d44eaabc"
         ),
 
         // MARK: - UnityAds
@@ -219,38 +143,17 @@ let package = Package(
         .target(
             name: CAS.UnityAds.target,
             dependencies: [
-                .target(name: CAS.UnityAds.binarySDK),
                 .target(name: CAS.UnityAds.binaryAdapter),
                 .target(name: baseTarget),
                 .target(name: CAS.IronSource.target),
+                .product(name: "UnityAds", package: "Unity-Ads-Swift-Package"),
             ],
             path: CAS.UnityAds.pathAdapter,
-            linkerSettings: [
-                .linkedFramework("AdSupport"),
-                .linkedFramework("AudioToolbox"),
-                .linkedFramework("AVFoundation"),
-                .linkedFramework("CoreGraphics"),
-                .linkedFramework("CoreTelephony"),
-                .linkedFramework("Foundation"),
-                .linkedFramework("QuartzCore"),
-                .linkedFramework("StoreKit"),
-                .linkedFramework("SystemConfiguration"),
-                .linkedFramework("WebKit"),
-                .linkedFramework("UIKit"),
-                .linkedFramework("AVFAudio"),
-                .linkedFramework("CoreFoundation"),
-                .linkedFramework("Network"),
-            ]
-        ),
-        .binaryTarget(
-            name: CAS.UnityAds.binarySDK,
-            url: "https://github.com/Unity-Technologies/unity-ads-ios/releases/download/4.18.1/UnityAds.zip",
-            checksum: "62cac3d0df3c8e0106f364f0e456f4edb42b6364dd9dc44b1d13985f70c2fbd1"
         ),
         .binaryTarget(
             name: CAS.UnityAds.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.3/CASMediationUnityAds-4.18.1.0.zip",
-            checksum: "f38e157ece3e812b4ae5be722406f87488b5d755309b53333b1b99f3292b0b22"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationUnityAds-4.19.0.1.zip",
+            checksum: "dad7e7a46faebd3879d48908fc852059d9f084a6fea82075c3888c5c06fd244c"
         ),
 
         // MARK: HyprMX
@@ -258,16 +161,11 @@ let package = Package(
         .target(
             name: CAS.HyprMX.target,
             dependencies: [
-                .target(name: CAS.HyprMX.binarySDK),
                 .target(name: CAS.HyprMX.binaryAdapter),
                 .target(name: baseTarget),
+                .product(name: "HyprMX", package: "HyprMX-SDK-SPM"),
             ],
             path: CAS.HyprMX.pathAdapter
-        ),
-        .binaryTarget(
-            name: CAS.HyprMX.binarySDK,
-            url: "https://s3.amazonaws.com/prd-mobile-sdk-files/HyprMX/6.4.5-rcs/HyprMX_iOS_v6_4_5-b353.zip",
-            checksum: "12757ba6a0d357c2d6d4fce74ce003d9f73cc1af82443d35a094b8b26aa44bc0"
         ),
         .binaryTarget(
             name: CAS.HyprMX.binaryAdapter,
@@ -328,8 +226,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.AppLovin.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.0-beta1/CASMediationAppLovin-13.6.3.1.zip",
-            checksum: "aeb66ab705c7ea1e08f199e73863c1033b2416c2ce8b120d2eca34bd94bcdae5"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationAppLovin-13.6.3.2.zip",
+            checksum: "207fc857b2b0b96d94b677a13117849284602efd3af399c0dd9a6090f0c6ae0a"
         ),
 
         // MARK: GoogleAds
@@ -340,17 +238,14 @@ let package = Package(
                 .target(name: CAS.GoogleAds.binaryAdapter),
                 .target(name: baseTarget),
                 .target(name: CAS.IronSource.target),
-                .product(
-                    name: "GoogleMobileAds",
-                    package: "swift-package-manager-google-mobile-ads"
-                ),
+                .product(name: "GoogleMobileAds", package: "swift-package-manager-google-mobile-ads"),
             ],
             path: CAS.GoogleAds.pathAdapter
         ),
         .binaryTarget(
             name: CAS.GoogleAds.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.0-beta1/CASMediationGoogleAds-13.6.0.0.zip",
-            checksum: "69c29ea852e5020b1286229320ccdfcb94a87bb345c441dc45973ae1a114ac4f"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationGoogleAds-13.7.0.0.zip",
+            checksum: "610d12085cccd7cf265ddced18b1cd2f4fcfc4ff7028f6ff96d13aac294af1a9"
         ),
 
         // MARK: LiftoffMonetize
@@ -366,8 +261,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.LiftoffMonetize.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.4/CASMediationLiftoffMonetize-7.7.4.0.zip",
-            checksum: "24a4c2857d7cb301bf5dca772bfd9b1b3635c8e8cb1d03b13ec9a53f2553ab33"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationLiftoffMonetize-7.7.6.1.zip",
+            checksum: "537469a39a46faa98fbd7aba3f5cec253111b6eeb11c8917aabb5519fe83afb5"
         ),
 
         // MARK: - StartIO
@@ -383,8 +278,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.StartIO.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.6.6/CASMediationStartIO-4.13.1.0.zip",
-            checksum: "dc6f02dd0b14836ddc0562db69421f4d0661c5bcd72eaf7a386fd43f9b9d43a0"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationStartIO-4.14.0.0.zip",
+            checksum: "3aee950b882bb0579e7ec5ac178803c46609bed39abeba44b6e69e812096ec86"
         ),
 
         // MARK: - Mintegral
@@ -400,8 +295,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.Mintegral.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.4/CASMediationMintegral-8.1.5.0.zip",
-            checksum: "0cc7f53cdce8df19499f5eeb008a6a5efda16c5c14bb7ab063c8dc26181170b8"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationMintegral-8.1.6.0.zip",
+            checksum: "8034c1ce46d691e341ad5036f5ce45295d5c825629207addd1578fbbfa4e15f6"
         ),
 
         // MARK: - DTExchange
@@ -409,22 +304,17 @@ let package = Package(
         .target(
             name: CAS.DTExchange.target,
             dependencies: [
-                .target(name: CAS.DTExchange.binarySDK),
                 .target(name: CAS.DTExchange.binaryAdapter),
                 .target(name: baseTarget),
                 .target(name: CAS.IronSource.target),
+                .product(name: "DTExchangeSDK", package: "DTExchangeSDK-iOS-SPM"),
             ],
             path: CAS.DTExchange.pathAdapter
         ),
         .binaryTarget(
-            name: CAS.DTExchange.binarySDK,
-            url: "https://github.com/inner-active/InneractiveAdSDK-iOS/archive/refs/heads/8.4.7.zip",
-            checksum: "5e770df2125adee8596cef35ee7f7d52cf0865ac9f65df7b8ef4eaf2d9bba138"
-        ),
-        .binaryTarget(
             name: CAS.DTExchange.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.6.6/CASMediationDTExchange-8.4.7.0.zip",
-            checksum: "a5dedd40ff937ec8b4d16866ce5cfd75330f108c1847390653f484101ce59054"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationDTExchange-8.4.9.0.zip",
+            checksum: "9880ea100d47d7bcf15cb40084274236de410fd030d657d9f79308cb93a8d0a8"
         ),
 
         // MARK: - CASExchange
@@ -467,8 +357,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.CrossPromo.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.0/CASMediationCrossPromo-4.1.0.1.zip",
-            checksum: "d7a6b7a88ed3b0dd8d3fe3944f5356ac2d039a1de47087d8bfe2f4d1595e8dc5"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationCrossPromo-4.8.0.0.zip",
+            checksum: "cb9123705a7f45aabc495435d202a0ef55280a0560df884d77be7c9c24f2fced"
         ),
 
         // MARK: - InMobi
@@ -476,9 +366,9 @@ let package = Package(
         .target(
             name: CAS.InMobi.target,
             dependencies: [
-                .target(name: CAS.InMobi.binarySDK),
                 .target(name: CAS.InMobi.binaryAdapter),
                 .target(name: baseTarget),
+                .product(name: "InMobiSDK", package: "InMobiSDK-Swift-Package"),
             ],
             path: CAS.InMobi.pathAdapter,
             linkerSettings: [
@@ -487,14 +377,9 @@ let package = Package(
             ]
         ),
         .binaryTarget(
-            name: CAS.InMobi.binarySDK,
-            url: "https://dl.inmobi.com/inmobi-sdk/IM/InMobi-iOS-SDK-11.3.0.zip",
-            checksum: "1dc4583ad8718845180903bfa6accc17c5248eca6d0aabbd43b9fb0dad1f58ce"
-        ),
-        .binaryTarget(
             name: CAS.InMobi.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.6.6/CASMediationInMobi-11.3.0.0.zip",
-            checksum: "882fafb6488784aed214e16a4555261318eee543182f00881fd45b2b61b09ff6"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationInMobi-11.4.0.0.zip",
+            checksum: "0e6c181ab23299b1f3fffced9addb45f32bf75d6744db233cec2204bbec1c5fa"
         ),
 
         // MARK: - Yango Ads
@@ -502,29 +387,16 @@ let package = Package(
         .target(
             name: CAS.YangoAds.target,
             dependencies: [
-                .target(name: CAS.YangoAds.binarySDK),
                 .target(name: CAS.YangoAds.binaryAdapter),
-                .target(name: baseTarget),
                 .target(name: CAS.IronSource.target),
-                .product(name: "AppMetricaCore", package: "appmetrica-sdk-ios"),
-                .product(name: "AppMetricaLibraryAdapter", package: "appmetrica-sdk-ios"),
-                .product(name: "AppMetricaAdSupport", package: "appmetrica-sdk-ios"),
-                .product(name: "AppMetricaIDSync", package: "appmetrica-sdk-ios"),
+                .product(name: "YandexMobileAds", package: "yandex-ads-sdk-ios"),
             ],
             path: CAS.YangoAds.pathAdapter,
-            resources: [
-                .process("Resources")
-            ]
-        ),
-        .binaryTarget(
-            name: CAS.YangoAds.binarySDK,
-            url: "https://ads-mobile-sdk.s3.yandex.net/Yandex/YandexMobileAds/8.1.0/spm/ca020bc8-a791-449d-a228-e726ba619562.zip",
-            checksum: "796f41ce5f415e1e4281fe9d902ec4e3ea4f083013a910b5f7435733ac2decbc"
         ),
         .binaryTarget(
             name: CAS.YangoAds.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.4/CASMediationYangoAds-8.1.0.1.zip",
-            checksum: "5da3fa593b43a94a1a88f0e885491f1a3da185ff3af9e3226eeb197f0ca76736"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationYangoAds-8.3.0.0.zip",
+            checksum: "c2771744b19e0a3de11d08746acef3b88fc5169ced75133c2d79a625bd5cbfd5"
         ),
 
         // MARK: - YsoNetwork
@@ -532,23 +404,16 @@ let package = Package(
         .target(
             name: CAS.YsoNetwork.target,
             dependencies: [
-                .target(name: CAS.YsoNetwork.binarySDK),
                 .target(name: CAS.YsoNetwork.binaryAdapter),
                 .target(name: baseTarget),
+                .product(name: "YsoNetwork", package: "ysonetwork-ios-sdk"),
             ],
             path: CAS.YsoNetwork.pathAdapter
         ),
         .binaryTarget(
-            name: CAS.YsoNetwork.binarySDK,
-            url:
-                "https://bitbucket.org/ysocorp/ysonetwork-ios-sdk/get/794e6daf0ed2dc2f237714f4e86d24de7a888a63.zip",
-            checksum: "237087e30b6382f0ff11b53b2c892f213d0fb34e6bf5aec3810316a3728dc978"
-        ),
-        .binaryTarget(
             name: CAS.YsoNetwork.binaryAdapter,
-            url:
-                "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.2.1/CASMediationYsoNetwork-1.1.31.2.zip",
-            checksum: "a8b70539b115757e8784fbc0109884e6c7451ca2d50c3bcb03e858e96921c561"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationYsoNetwork-1.2.1.0.zip",
+            checksum: "c2f3d905eb2ab92bd4b4d6dc34c4a5e641bfdeddae40e6432c85baf95f1fd810"
         ),
 
         // MARK: - AudienceNetwork
@@ -556,46 +421,16 @@ let package = Package(
         .target(
             name: CAS.AudienceNetwork.target,
             dependencies: [
-                .target(name: CAS.AudienceNetwork.binarySDK),
                 .target(name: CAS.AudienceNetwork.binaryAdapter),
-                .target(name: baseTarget),
-                .target(name: CAS.IronSource.target),
+                .target(name: CAS.GoogleAds.target),
+                .product(name: "FBAudienceNetwork", package: "FBAudienceNetwork"),
             ],
             path: CAS.AudienceNetwork.pathAdapter,
-            linkerSettings: [
-                .linkedFramework("AudioToolbox"),
-                .linkedFramework("AppTrackingTransparency"),
-                .linkedFramework("StoreKit"),
-                .linkedFramework("CoreGraphics"),
-                .linkedFramework("UIKit"),
-                .linkedFramework("Foundation"),
-                .linkedFramework("Security"),
-                .linkedFramework("CoreImage"),
-                .linkedFramework("AVFoundation"),
-                .linkedFramework("CoreMedia"),
-                .linkedFramework("AdSupport"),
-                .linkedFramework("CFNetwork"),
-                .linkedFramework("CoreMotion"),
-                .linkedFramework("CoreTelephony"),
-                .linkedFramework("LocalAuthentication"),
-                .linkedFramework("SafariServices"),
-                .linkedFramework("SystemConfiguration"),
-                .linkedFramework("VideoToolbox"),
-                .linkedFramework("WebKit"),
-                .linkedLibrary("c++"),
-                .linkedLibrary("xml2"),
-                .linkedLibrary("z"),
-            ]
-        ),
-        .binaryTarget(
-            name: CAS.AudienceNetwork.binarySDK,
-            url: "https://developers.facebook.com/resources/FBAudienceNetwork-6.21.1.zip",
-            checksum: "acb53ced101d439131f0c5d0cf3da609129227140a66e4f51ec77a368a753843"
         ),
         .binaryTarget(
             name: CAS.AudienceNetwork.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.6.3/CASMediationAudienceNetwork-6.21.1.0.zip",
-            checksum: "44b2b069d960c79efeda502fcbf839ebdd5d23ad859c8a4a82f580efd7f38925"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationAudienceNetwork-6.22.0.0.zip",
+            checksum: "5f0f786a4635030eb04335eb9c6756e7a73179e68728064dfadc6f9955df07f7"
         ),
 
         // MARK: - Pangle
@@ -611,8 +446,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.Pangle.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.0-beta1/CASMediationPangle-8.1.1.0.0.zip",
-            checksum: "02576f04030a9e40ebce1b8af118d6fd92c73c975376faed3f6d4d445e5e44b6"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationPangle-8.2.0.8.0.zip",
+            checksum: "1035c60dcf805e7627751ee28c6b4060caaa4b471e4cb961bf4562663c365928"
         ),
 
         // MARK: - Maticoo
@@ -644,15 +479,15 @@ let package = Package(
             name: CAS.PubMatic.target,
             dependencies: [
                 .target(name: CAS.PubMatic.binaryAdapter),
-                .target(name: CAS.AppLovin.target),
+                .target(name: baseTarget),
                 .product(name: "OpenWrapSDK", package: "OpenWrapSDK-Swift-Package"),
             ],
             path: CAS.PubMatic.pathAdapter
         ),
         .binaryTarget(
             name: CAS.PubMatic.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.1/CASMediationPubMatic-5.1.1.0.zip",
-            checksum: "0a48379ab6604bd4f688fdab37f278008e1ee392f960d70e636faf48802d3dfe"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationPubMatic-5.2.0.1.zip",
+            checksum: "af0e3dfe44dfc348b4b9b623f59ac7f6cfeb8b1d9d7f7ba7a4064065dd87a99b"
         ),
 
         // MARK: - Ogury
@@ -677,37 +512,16 @@ let package = Package(
         .target(
             name: CAS.Chartboost.target,
             dependencies: [
-                .target(name: CAS.Chartboost.binarySDK),
                 .target(name: CAS.Chartboost.binaryAdapter),
-                .target(name: baseTarget),
                 .target(name: CAS.IronSource.target),
+                .product(name: "ChartboostSDK", package: "chartboost-monetization-ios-sdk"),
             ],
             path: CAS.Chartboost.pathAdapter,
-            linkerSettings: [
-                .linkedFramework("AdSupport"),
-                .linkedFramework("CoreGraphics"),
-                .linkedFramework("StoreKit"),
-                .linkedFramework("Foundation"),
-                .linkedFramework("UIKit"),
-                .linkedFramework("WebKit"),
-                .linkedFramework("AVFoundation"),
-                .linkedFramework("CoreMedia"),
-                .linkedFramework("SystemConfiguration"),
-                .linkedFramework("CoreFoundation"),
-                .linkedFramework("SafariServices"),
-                .linkedFramework("CoreTelephony"),
-                .linkedFramework("AppTrackingTransparency"),
-            ]
-        ),
-        .binaryTarget(
-            name: CAS.Chartboost.binarySDK,
-            url: "https://s3.amazonaws.com/chartboost/sdk/9.12.0/Chartboost-iOS-9.12.0.zip",
-            checksum: "8a611fed3d3e76be3faf13ffcfda6703c90abcbf415d610a6270d515dc9ae271"
         ),
         .binaryTarget(
             name: CAS.Chartboost.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.6.6/CASMediationChartboost-9.12.0.0.zip",
-            checksum: "dbe2da10e7afc1b77c899c1c71fa134e5658dc93edc784aad2b3f8dfcd42f372"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationChartboost-9.13.0.0.zip",
+            checksum: "89000499e3a46ae9acc3b7726c9527195a95f2a936c3fe0f4221b693a71ca940"
         ),
 
         // MARK: - Verve
@@ -723,8 +537,8 @@ let package = Package(
         ),
         .binaryTarget(
             name: CAS.Verve.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.7.0/CASMediationVerve-3.8.1.0.zip",
-            checksum: "eb43caca6f0777f11041ce1e0de12e8f9b8c8827447851fa4f01c08fe960f67d"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationVerve-3.9.0.1.zip",
+            checksum: "1c579ce7f9db4cd62791389626ab97a075acbdcca1f5d810066807a42080a290"
         ),
 
         // MARK: - Bigo Ads
@@ -759,15 +573,49 @@ let package = Package(
             name: CAS.DisplayIO.target,
             dependencies: [
                 .target(name: CAS.DisplayIO.binaryAdapter),
-                .target(name: baseTarget),
-                .product(name: "DIOSDK", package: "DIOSDK"),
+                .target(name: CAS.AudienceNetwork.target),
+                .product(name: "DIOSDK-WithoutFBAudienceNetwork", package: "DIOSDK"),
             ],
             path: CAS.DisplayIO.pathAdapter
         ),
         .binaryTarget(
             name: CAS.DisplayIO.binaryAdapter,
-            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.0-beta1/CASMediationDisplayIO-4.7.4.0.zip",
-            checksum: "b51c2109552732e36033a6b4f5e38550838bedb71b2e3c3becffc7db9569735d"
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationDisplayIO-4.7.5.0.zip",
+            checksum: "99244f3bc976f5ae1685c32f5e096184cd27b4083a281ddbd7269d14187bf32d"
+        ),
+
+        // MARK: - Smaato
+
+        .target(
+            name: CAS.Smaato.target,
+            dependencies: [
+                .target(name: CAS.Smaato.binaryAdapter),
+                .target(name: CAS.IronSource.target),
+                .product(name: "SmaatoSDK", package: "Smaato-ios-sdk-standalone"),
+            ],
+            path: CAS.Smaato.pathAdapter
+        ),
+        .binaryTarget(
+            name: CAS.Smaato.binaryAdapter,
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationSmaato-23.2.0.1.zip",
+            checksum: "7190ff69becb75e2d42b6fc50e357189c605f874894c6b2403742607bc96db5f"
+        ),
+
+        // MARK: - Moloco
+
+        .target(
+            name: CAS.Moloco.target,
+            dependencies: [
+                .target(name: CAS.Moloco.binaryAdapter),
+                .target(name: CAS.IronSource.target),
+                .product(name: "MolocoSDK", package: "moloco-sdk-ios-spm"),
+            ],
+            path: CAS.Moloco.pathAdapter
+        ),
+        .binaryTarget(
+            name: CAS.Moloco.binaryAdapter,
+            url: "https://github.com/cleveradssolutions/CAS-iOS/releases/download/4.8.1-beta1/CASMediationMoloco-4.9.0.0.zip",
+            checksum: "1d86f9f5e7256e3c1cae51f6a8766ca22ebb121ba936a566e2ebd603328c24ec"
         ),
 
     ]