+++ Adapters/AudienceNetwork/CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Meta Audience Network iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationAudienceNetwork', '6.21.1.0'
+pod 'CASMediationAudienceNetwork', '6.22.0.0'
 ```
+
+### 6.22.0.0
+- Certified with AudienceNetwork - 6.22.0
+- Replaced the dependency with the official SPM git.
+- Minimum supported iOS 15
 
 ### 6.21.1.0
 - Certified with AudienceNetwork - 6.21.1