+++ Adapters/DisplayIO/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## Display IO iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationDisplayIO', '4.7.4.0'
+pod 'CASMediationDisplayIO', '4.7.5.0'
 ```
+
+### 4.7.5.0
+- Certified with DisplayIO - 4.7.5
+- Fixed an SPM integration error caused by a duplicate Meta Audience Network SDK.
 
 ### 4.7.4.0
 - Certified with DisplayIO - 4.7.4