+++ Adapters/InMobi/CHANGELOG.md
@@ -1,7 +1,12 @@
 ## InMobi iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationInMobi', '11.3.0.0'
+pod 'CASMediationInMobi', '11.4.0.0'
 ```
+
+### 11.4.0.0
+- Certified with InMobi - 11.4.0
+- Added Adaptive banner size support.
+- Replaced the dependency with the official SPM git.
 
 ### 11.3.0.0
 - Certified with InMobi - 11.3.0