+++ Adapters/IronSource/CHANGELOG.md
@@ -1,7 +1,11 @@
 ## IronSource iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationIronSource', '9.4.2.0'
+pod 'CASMediationIronSource', '9.5.0.1'
 ```
+
+### 9.5.0.1
+- Certified with IronSource - 9.5.0
+- Replaced the dependency with the official SPM git.
 
 ### 9.4.2.0
 - Certified with IronSource - 9.4.2