+++ Adapters/Moloco/CHANGELOG.md
@@ -1,7 +1,13 @@
 ## Moloco iOS Mediation Adapter Changelog
 ```ruby
-pod 'CASMediationMoloco', '4.8.0.0'
+pod 'CASMediationMoloco', '4.9.0.0'
 ```
+
+### 4.9.0.0
+- Certified with Moloco - 4.9.0
+- Added Adaptive banner size support.
+- Added Swift Package Manager integration support.
+- The AppLovin dependency is no longer required.
 
 ### 4.8.0.0
 - Certified with Moloco - 4.8.0